refactor: clarify codebase structure boundaries
