+++ internal/app/input/on_plugin.go
@@ -1,4 +1,4 @@
-// Plugin selector: model, state, runtime, keymap, navigation, actions, commands, load, reset.
+// Plugin selector: navigation, actions, loading, and reset.
 package input
 
 import (
@@ -15,480 +15,6 @@ import (
 	"github.com/genai-io/san/internal/app/kit"
 	coreplugin "github.com/genai-io/san/internal/plugin"
 )
-
-// ── Types ────────────────────────────────────────────────────────────────────
-
-// pluginTab represents the active tab in the plugin selector
-type pluginTab int
-
-const (
-	pluginTabDiscover pluginTab = iota
-	pluginTabInstalled
-	pluginTabMarketplaces
-)
-
-// pluginLevel represents the navigation level within the plugin selector
-type pluginLevel int
-
-const (
-	pluginLevelTabList pluginLevel = iota
-	pluginLevelDetail
-	pluginLevelInstallOptions
-	pluginLevelAddMarketplace
-	pluginLevelBrowsePlugins
-)
-
-type pluginAction struct {
-	Label  string
-	Action string
-}
-
-// pluginItem represents a plugin in the selector
-type pluginItem struct {
-	Name        string
-	FullName    string
-	Description string
-	Version     string
-	Scope       coreplugin.Scope
-	Enabled     bool
-	Path        string
-	Skills      int
-	Agents      int
-	Commands    int
-	Hooks       int
-	MCP         int
-	LSP         int
-	Errors      []string
-	Author      string
-	Homepage    string
-	Marketplace string
-}
-
-// pluginDiscoverItem represents a plugin available in a marketplace
-type pluginDiscoverItem struct {
-	Name        string
-	Description string
-	Marketplace string
-	Author      string
-	Installed   bool
-	Homepage    string
-	Version     string
-}
-
-// pluginMarketplaceItem represents a marketplace in the selector
-type pluginMarketplaceItem struct {
-	ID          string
-	Source      string
-	SourceType  string
-	Available   int
-	Installed   int
-	LastUpdated string
-	IsOfficial  bool
-}
-
-// PluginSelector holds state for the plugin selector
-type PluginSelector struct {
-	registry *coreplugin.Registry
-
-	active         bool
-	width          int
-	height         int
-	lastMessage    string
-	isError        bool
-	maxVisible     int
-	isLoading      bool
-	loadingMsg     string
-	loadingFrame   int
-	loadingTicking bool
-
-	activeTab pluginTab
-
-	installedPlugins  map[coreplugin.Scope][]pluginItem
-	installedScopes   []coreplugin.Scope
-	installedFlatList []pluginItem
-	discoverPlugins   []pluginDiscoverItem
-	marketplaces      []pluginMarketplaceItem
-
-	level        pluginLevel
-	selectedIdx  int
-	scrollOffset int
-	detailScroll int
-
-	searchQuery   string
-	filteredItems []any
-
-	detailPlugin      *pluginItem
-	detailDiscover    *pluginDiscoverItem
-	detailMarketplace *pluginMarketplaceItem
-	actions           []pluginAction
-	actionIdx         int
-	parentIdx         int
-
-	addMarketplaceInput string
-	addDialogCursor     int
-
-	browseMarketplaceID string
-	browsePlugins       []pluginDiscoverItem
-
-	marketplaceManager *coreplugin.MarketplaceManager
-	installer          *coreplugin.Installer
-}
-
-// Plugin intent messages — each requests one async mutating operation. They
-// bridge from keybindings / executeAction into UpdatePlugin, where deps (cwd,
-// reload) and the spinner live.
-type pluginEnableMsg struct{ PluginName string }
-type pluginDisableMsg struct{ PluginName string }
-type pluginUninstallMsg struct{ PluginName string }
-type pluginMarketplaceRemoveMsg struct{ ID string }
-
-type pluginInstallMsg struct {
-	PluginName  string
-	Marketplace string
-	Scope       coreplugin.Scope
-}
-
-// pluginOpResultMsg is the unified outcome of every async mutating operation
-// (install / sync / remove / uninstall / enable / disable). One spinner path,
-// one result handler — see runPluginOp and the pluginOpResultMsg case.
-type pluginOpResultMsg struct {
-	okMsg  string // success status, e.g. "Installed foo"
-	errCtx string // failure verb + subject, e.g. "install foo"
-	err    error
-	pop    bool // goBack() on success (leave the detail / options view)
-	reload bool // ReloadAfterPluginChange() on success
-}
-
-// NewPluginSelector creates a new PluginSelector
-func NewPluginSelector(reg *coreplugin.Registry) PluginSelector {
-	cwd, _ := os.Getwd()
-	return PluginSelector{
-		registry:           reg,
-		active:             false,
-		maxVisible:         15,
-		activeTab:          pluginTabInstalled,
-		installedPlugins:   make(map[coreplugin.Scope][]pluginItem),
-		marketplaceManager: coreplugin.NewMarketplaceManager(cwd),
-		installer:          coreplugin.NewInstaller(reg, cwd),
-	}
-}
-
-// IsActive returns whether the selector is active
-func (s *PluginSelector) IsActive() bool {
-	return s.active
-}
-
-// ── Runtime ──────────────────────────────────────────────────────────────────
-
-// UpdatePlugin routes plugin management messages.
-func UpdatePlugin(deps OverlayDeps, state *PluginSelector, msg tea.Msg) (tea.Cmd, bool) {
-	switch msg := msg.(type) {
-	case pluginEnableMsg:
-		reg, name := state.registry, msg.PluginName
-		return tea.Batch(
-			state.beginLoading("Enabling "+name+"..."),
-			runPluginOp(pluginOpResultMsg{okMsg: "Enabled " + name, errCtx: "enable " + name},
-				func() error { return reg.Enable(name, coreplugin.ScopeUser) }),
-		), true
-
-	case pluginDisableMsg:
-		reg, name := state.registry, msg.PluginName
-		return tea.Batch(
-			state.beginLoading("Disabling "+name+"..."),
-			runPluginOp(pluginOpResultMsg{okMsg: "Disabled " + name, errCtx: "disable " + name},
-				func() error { return reg.Disable(name, coreplugin.ScopeUser) }),
-		), true
-
-	case pluginInstallMsg:
-		reg, cwd := state.registry, deps.Cwd
-		name, market, scope := msg.PluginName, msg.Marketplace, msg.Scope
-		return tea.Batch(
-			state.beginLoading("Installing "+name+"..."),
-			runPluginOp(pluginOpResultMsg{okMsg: "Installed " + name, errCtx: "install " + name, pop: true, reload: true},
-				func() error {
-					ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
-					defer cancel()
-					return coreplugin.Install(ctx, reg, cwd, coreplugin.FormatPluginRef(name, market), scope)
-				}),
-		), true
-
-	case pluginUninstallMsg:
-		inst, name := state.installer, msg.PluginName
-		return tea.Batch(
-			state.beginLoading("Uninstalling "+name+"..."),
-			runPluginOp(pluginOpResultMsg{okMsg: "Uninstalled " + name, errCtx: "uninstall " + name, pop: true, reload: true},
-				func() error { return inst.Uninstall(name, coreplugin.ScopeUser) }),
-		), true
-
-	case pluginMarketplaceRemoveMsg:
-		mgr, id := state.marketplaceManager, msg.ID
-		return tea.Batch(
-			state.beginLoading("Removing "+id+"..."),
-			runPluginOp(pluginOpResultMsg{okMsg: "Removed " + id, errCtx: "remove " + id, pop: true},
-				func() error { return mgr.Remove(id) }),
-		), true
-
-	case pluginOpResultMsg:
-		state.endLoading()
-		if msg.err != nil {
-			state.setError(fmt.Sprintf("Failed to %s: %v", msg.errCtx, msg.err))
-		} else {
-			state.setSuccess(msg.okMsg)
-			if msg.pop {
-				state.goBack()
-			}
-		}
-		state.refreshAfterOp()
-		if msg.err == nil && msg.reload {
-			_ = deps.ReloadAfterPluginChange()
-		}
-		return nil, true
-
-	case pluginLoadingTickMsg:
-		if !state.isLoading {
-			state.loadingTicking = false
-			return nil, true
-		}
-		state.loadingFrame++
-		return pluginLoadingTick(), true
-	}
-	return nil, false
-}
-
-type pluginLoadingTickMsg struct{}
-
-func pluginLoadingTick() tea.Cmd {
-	return tea.Tick(80*time.Millisecond, func(time.Time) tea.Msg {
-		return pluginLoadingTickMsg{}
-	})
-}
-
-// beginLoading marks the selector as loading with msg and returns a tick
-// cmd to drive the spinner — or nil if a tick is already in flight, so
-// overlapping loading actions don't double the spinner cadence.
-func (s *PluginSelector) beginLoading(msg string) tea.Cmd {
-	s.isLoading = true
-	s.loadingMsg = msg
-	return s.startLoadingTick()
-}
-
-// startLoadingTick returns a tick cmd only if no tick is already in flight.
-func (s *PluginSelector) startLoadingTick() tea.Cmd {
-	if s.loadingTicking {
-		return nil
-	}
-	s.loadingTicking = true
-	return pluginLoadingTick()
-}
-
-// endLoading clears the spinner. The in-flight tick stops itself on the next
-// frame once it sees isLoading false.
-func (s *PluginSelector) endLoading() {
-	s.isLoading = false
-	s.loadingMsg = ""
-}
-
-// runPluginOp wraps a blocking mutation in a command that reports the unified
-// pluginOpResultMsg. Pair it with beginLoading so the spinner runs until the
-// result lands.
-func runPluginOp(res pluginOpResultMsg, work func() error) tea.Cmd {
-	return func() tea.Msg {
-		res.err = work()
-		return res
-	}
-}
-
-// ── Keymap ───────────────────────────────────────────────────────────────────
-
-// HandleKeypress handles a keypress and returns a command if needed.
-func (s *PluginSelector) HandleKeypress(key tea.KeyMsg) tea.Cmd {
-	if s.level == pluginLevelAddMarketplace {
-		return s.handleAddMarketplaceKeypress(key)
-	}
-	if s.level == pluginLevelDetail || s.level == pluginLevelInstallOptions {
-		return s.handleDetailKeypress(key)
-	}
-	if s.level == pluginLevelBrowsePlugins {
-		return s.handleBrowseKeypress(key)
-	}
-	return s.handleListKeypress(key)
-}
-
-// HandlePaste appends bracketed-paste content to the marketplace-source field.
-// A source is a single token (URL, owner/repo, or local path), so embedded
-// newlines and surrounding whitespace are stripped rather than inserted.
-func (s *PluginSelector) HandlePaste(content string) tea.Cmd {
-	if s.level != pluginLevelAddMarketplace {
-		return nil
-	}
-	content = strings.NewReplacer("\r", "", "\n", "").Replace(content)
-	content = strings.TrimSpace(content)
-	if content == "" {
-		return nil
-	}
-	s.addMarketplaceInput += content
-	s.clearMessage()
-	return nil
-}
-
-func (s *PluginSelector) handleAddMarketplaceKeypress(key tea.KeyMsg) tea.Cmd {
-	switch key.String() {
-	case "esc":
-		s.goBack()
-		return nil
-	case "enter":
-		return s.addMarketplace()
-	case "backspace":
-		if len(s.addMarketplaceInput) > 0 {
-			s.addMarketplaceInput = s.addMarketplaceInput[:len(s.addMarketplaceInput)-1]
-			s.clearMessage()
-		}
-		return nil
-	default:
-		// Typed text capture: append printable characters to the input.
-		if text := key.Key().Text; text != "" {
-			s.addMarketplaceInput += text
-			s.clearMessage()
-		}
-		return nil
-	}
-}
-
-func (s *PluginSelector) handleDetailKeypress(key tea.KeyMsg) tea.Cmd {
-	if s.handleNavigationKey(key, true) {
-		return nil
-	}
-	switch key.String() {
-	case "enter":
-		return s.executeAction()
-	case "esc", "left", "h":
-		s.goBack()
-	}
-	return nil
-}
-
-func (s *PluginSelector) handleBrowseKeypress(key tea.KeyMsg) tea.Cmd {
-	if s.handleNavigationKey(key, true) {
-		return nil
-	}
-	switch key.String() {
-	case "enter":
-		if s.selectedIdx < len(s.browsePlugins) {
-			p := s.browsePlugins[s.selectedIdx]
-			s.detailDiscover = &p
-			s.actions = s.buildDiscoverActions(p)
-			s.actionIdx = 0
-			s.level = pluginLevelDetail
-		}
-	case "esc", "left":
-		s.goBack()
-	}
-	return nil
-}
-
-// handleNavigationKey handles common up/down navigation keys, returns true if handled.
-func (s *PluginSelector) handleNavigationKey(key tea.KeyMsg, vimKeys bool) bool {
-	switch key.String() {
-	case "up", "ctrl+p":
-		s.MoveUp()
-		return true
-	case "down", "ctrl+n":
-		s.MoveDown()
-		return true
-	case "k":
-		if vimKeys {
-			s.MoveUp()
-			return true
-		}
-	case "j":
-		if vimKeys {
-			s.MoveDown()
-			return true
-		}
-	}
-	return false
-}
-
-func (s *PluginSelector) handleListKeypress(key tea.KeyMsg) tea.Cmd {
-	if s.searchQuery == "" {
-		switch key.String() {
-		case "tab", "right":
-			s.NextTab()
-			return nil
-		case "shift+tab", "left":
-			s.PrevTab()
-			return nil
-		}
-	}
-
-	if s.handleNavigationKey(key, s.searchQuery == "") {
-		return nil
-	}
-
-	switch key.String() {
-	case "enter":
-		s.enterDetail()
-		return nil
-	case "esc":
-		if s.searchQuery != "" {
-			s.searchQuery = ""
-			s.updateFilter()
-			return nil
-		}
-		s.Cancel()
-		return func() tea.Msg { return kit.DismissedMsg{} }
-	case "backspace":
-		if len(s.searchQuery) > 0 {
-			s.searchQuery = s.searchQuery[:len(s.searchQuery)-1]
-			s.updateFilter()
-		}
-		return nil
-	default:
-		// Typed text capture (printable runes drive vim keys / search).
-		if text := key.Key().Text; text != "" {
-			return s.handleListRuneKey(text)
-		}
-	}
-	return nil
-}
-
-// handleListRuneKey handles rune key input in list view.
-func (s *PluginSelector) handleListRuneKey(r string) tea.Cmd {
-	if s.searchQuery == "" {
-		switch r {
-		case "l":
-			s.enterDetail()
-			return nil
-		case " ":
-			return s.toggleSelectedPlugin()
-		case "u":
-			return s.handleMarketplaceAction(func(m pluginMarketplaceItem) tea.Cmd {
-				return s.syncMarketplace(m.ID)
-			})
-		case "r":
-			return s.handleMarketplaceAction(func(m pluginMarketplaceItem) tea.Cmd {
-				return func() tea.Msg { return pluginMarketplaceRemoveMsg{ID: m.ID} }
-			})
-		}
-	}
-	s.searchQuery += r
-	s.updateFilter()
-	return nil
-}
-
-// handleMarketplaceAction executes an action on the selected marketplace.
-func (s *PluginSelector) handleMarketplaceAction(action func(pluginMarketplaceItem) tea.Cmd) tea.Cmd {
-	if s.activeTab != pluginTabMarketplaces || s.selectedIdx == 0 {
-		return nil
-	}
-	mktIdx := s.selectedIdx - 1
-	if mktIdx < len(s.filteredItems) {
-		if m, ok := s.filteredItems[mktIdx].(pluginMarketplaceItem); ok {
-			return action(m)
-		}
-	}
-	return nil
-}
 
 // ── Navigation ───────────────────────────────────────────────────────────────
